# Use `Object.assign` instead of custom `copy` function