While investigating possible ways of mitigating the build time costs identified in emberjs/ember.js#19750, we noticed this custom `copy` implementation. Since it performs the same basic behavior as `Object.assign`, I replaced it with `Object.assign`.

This has no impact on runtime performance (see below), and drops about 1.1KB of non-minified, uncompressed bytes across the wire.

## Runtime Performance

I ran it against the set of templates we pulled together in https://github.com/brendenpalmer/repro-ember-3-25-template-regression, using [this PR](https://github.com/brendenpalmer/repro-ember-3-25-template-regression/pull/2) with the following results.

While it looks like a small percentage decrease against the baseline, the variance between runs is significant enough (even with 5 runs per compiler) that I don't think this is meaningful. That said, I'm still going to run this change through our app as well, and if that looks more meaningful I'll report that below.

### With current Glimmer VM `master`

|                            source                             |  time   |
| ------------------------------------------------------------- | ------- |
| ember-source@3.24.5 (ember-source-3-24)                       | 1958.2  |
| ember-source@3.25.4 (ember-source-3-25)                       | 12742.2 |
| ember-source@3.28.1 (ember-source-3-28)                       | 12499.4 |
| @glimmer/compiler@0.65.3 (glimmer-compiler-ember-source-3-24) | 1981.2  |
| @glimmer/compiler@0.80.0 (glimmer-compiler-ember-source-3-28) | 13433.6 |
| @glimmer/compiler@0.80.0 (glimmer-compiler-experiment)        | 4201.4  |

### With `Object.assign`

This adds in `Object.assign` replacing `copy`.

|                            source                             |  time   |
| ------------------------------------------------------------- | ------- |
| ember-source@3.24.5 (ember-source-3-24)                       | 2147.2  |
| ember-source@3.25.4 (ember-source-3-25)                       | 19060.6 |
| ember-source@3.28.1 (ember-source-3-28)                       | 15271.6 |
| @glimmer/compiler@0.65.3 (glimmer-compiler-ember-source-3-24) | 2246    |
| @glimmer/compiler@0.80.0 (glimmer-compiler-ember-source-3-28) | 21358   |
| @glimmer/compiler@0.80.0 (glimmer-compiler-experiment)        | 4369    |