@chriskrycho 2021-09-21
This is failing b/c the repo hard-errors against use of `Object.assign`:

> Error: Unexpected use of Object.assign. Object.assign cannot be used because it is unavailable in IE11. This error was likely caused by using syntax like the spread operator ({ ...obj }) which transpiles to Object.assign. Please use alternate syntax that is compatible with IE11.

It's not clear to me why we don't just let users configure an `Object.assign` polyfill if needed?