@chriskrycho 2021-09-22
Addendum on the above comment: Ember's own build process *does* polyfill this correctly anyway, so it's unclear why Glimmer's own test suite is forbidding it.